requirements.txt
Click==7.0
Flask==1.1.1
loguru==0.3.2
numpy==1.16.4
pandas==0.25.0
pymongo==3.8.0
textblob==0.15.3
tweepy==3.8.0
python-decouple==3.1
